# openapi-fastapi.yaml
openapi: 3.0.3
info:
  title: SKAX Physical Risk AI Agent API (FastAPI)
  version: 1.0.0
  description: >
    사업장 기후 물리적 리스크 AI Agent 분석 시스템 API

    Spring Boot 서버로부터 요청을 받아 AI Agent를 통해 기후 리스크를 분석하고 결과를 반환합니다.

    주요 기능:
    - AI Agent 기반 물리적 리스크 분석
    - 비동기 작업 처리 및 상태 관리
    - 과거 재난 이력 분석
    - SSP 시나리오별 미래 리스크 전망
    - 재무 영향 분석 (AAL, CVaR)
    - 취약성 평가
    - 사업장 이전 시뮬레이션
    - 기후 변화 시뮬레이션
servers:
  - url: https://ai-api.skax.co.kr
    description: Production
  - url: https://staging-ai-api.skax.co.kr
    description: Staging
  - url: http://localhost:8000
    description: Local Development
...
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Spring Boot 서버 인증용 API Key
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: ANALYSIS_FAILED
        message:
          type: string
          example: 분석 중 오류가 발생했습니다.
        detail:
          type: string
          description: 상세 오류 메시지
        timestamp:
          type: string
          format: date-time

    # ==========================================
    # Site Input Schemas
    # ==========================================
    SiteInfo:
      type: object
      required: [id, name, latitude, longitude, industry]
      properties:
        id:
          type: string
          format: uuid
          description: 사업장 ID
        name:
          type: string
          example: 대덕 데이터센터
        address:
          type: string
          example: 대전광역시 유성구 대덕대로 512
        city:
          type: string
          example: 대전광역시
        latitude:
          type: number
          format: double
          example: 36.368
        longitude:
          type: number
          format: double
          example: 127.364
        industry:
          type: string
          example: DC(Data Center)
          description: 업종
        buildingAge:
          type: integer
          description: 건물 연령 (년)
          example: 15
        buildingType:
          type: string
          description: 건물 유형
          example: 철근 콘크리트
        seismicDesign:
          type: boolean
          description: 내진설계 여부
          default: false
        floorArea:
          type: number
          format: double
          description: 연면적 (m²)
        assetValue:
          type: number
          format: double
          description: 자산 가치 (KRW)
        employeeCount:
          type: integer
          description: 직원 수

    # ==========================================
    # Analysis Request/Response Schemas
    # ==========================================
    StartAnalysisRequest:
      type: object
      required: [site]
      properties:
        site:
          $ref: '#/components/schemas/SiteInfo'
        hazardTypes:
          type: array
          description: 분석할 리스크 유형 (미지정시 전체)
          items:
            type: string
            enum: [폭염, 한파, 태풍, 홍수, 가뭄, 산불, 해안침수, 도시침수, 물부족]
          example: [폭염, 태풍, 홍수]
        priority:
          type: string
          enum: [low, normal, high]
          default: normal
          description: 분석 우선순위
        options:
          type: object
          description: 분석 옵션
          properties:
            includeFinancialImpact:
              type: boolean
              default: true
            includeVulnerability:
              type: boolean
              default: true
            includePastEvents:
              type: boolean
              default: true
            sspScenarios:
              type: array
              items:
                type: string
                enum: [SSP1-2.6, SSP2-4.5, SSP3-7.0, SSP5-8.5]
              default: [SSP2-4.5, SSP5-8.5]
    AnalysisJobStatus:
      type: object
      properties:
        jobId:
          type: string
          format: uuid
          description: 작업 ID
        siteId:
          type: string
          format: uuid
          description: 사업장 ID
        status:
          type: string
          enum: [queued, running, completed, failed]
          description: 작업 상태
        progress:
          type: integer
          minimum: 0
          maximum: 100
          description: 진행률 (%)
        currentNode:
          type: string
          example: physical_risk_score
          description: 현재 실행 중인 Agent Node
        currentHazard:
          type: string
          example: 폭염
          description: 현재 분석 중인 Hazard
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        estimatedCompletionTime:
          type: string
          format: date-time
          description: 예상 완료 시간
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            node:
              type: string
              description: 오류 발생 Node
...
    AnalysisOverviewResponse:
      type: object
      description: 사업장 리스크 분석 개요
      properties:
        siteId:
          type: string
          format: uuid
        analyzedAt:
          type: string
          format: date-time
        mainHazard:
          type: string
          example: 폭염
          description: 가장 높은 물리적 리스크 종류
        mainHazardScore:
          type: integer
          minimum: 0
          maximum: 100
          description: 주요 위험요인 물리적 리스크 점수
        overallRiskScore:
          type: integer
          minimum: 0
          maximum: 100
          description: 종합 리스크 점수
        riskLevel:
          type: string
          enum: [낮음, 보통, 높음, 심각]
          description: 리스크 레벨
        aal:
          type: number
          format: float
          description: 연평균 자산 손실률 (AAL, 0.0-1.0)
          example: 0.023
        annualExpectedLoss:
          type: number
          format: double
          description: 연평균 재무 손실액 (KRW) = AAL × 자산가치
          example: 230000000
        assetValue:
          type: number
          format: double
          description: 자산 가치 (KRW)
          example: 10000000000
        recommendations:
          type: array
          description: AI가 생성한 추천사항
          items:
            type: string
          example:
            - 냉각 시스템 용량 증설 검토
            - 폭염 대비 비상 전력 공급 계획 수립

    # ==========================================
    # Physical Risk Score Schemas
    # ==========================================
    PhysicalRiskScoreResponse:
      type: object
      description: Hazard별 물리적 리스크 점수
      properties:
        siteId:
          type: string
          format: uuid
        analyzedAt:
          type: string
          format: date-time
        hazardType:
          type: string
          example: 폭염
          description: 조회한 위험요인 (전체인 경우 null)
        physicalRiskScores:
          type: array
          description: Hazard별 물리적 리스크 점수
          items:
            $ref: '#/components/schemas/PhysicalRiskBarItem'

    # ==========================================
    # Past Events Schemas
    # ==========================================
...
    PastEventsResponse:
      type: object
      properties:
        siteId:
          type: string
          format: uuid
        summary:
          $ref: '#/components/schemas/PastEventSummary'
        events:
          type: array
          items:
            $ref: '#/components/schemas/PastEventItem'

    # ==========================================
    # SSP Projection Schemas
    # ==========================================
...
    SSPProjectionResponse:
      type: object
      properties:
        siteId:
          type: string
          format: uuid
        hazardType:
          type: string
          example: 폭염
        timeScale:
          type: string
          example: 장기
          enum: [단기, 중기, 장기]
        projections:
          type: array
          description: 시나리오별 전망
          items:
            $ref: '#/components/schemas/SSPScenarioProjection'
        adaptationActions:
          type: array
          description: AI가 제안하는 적응 조치
          items:
            type: object
            properties:
              decade:
                type: string
                example: "2030s"
              action:
                type: string
                example: 냉각 시스템 용량 50% 증설
              priority:
                type: string
                enum: [low, medium, high, critical]
              estimatedCost:
                type: number
                description: 예상 비용 (KRW)
        confidenceLevel:
          type: string
          enum: [low, medium, high]
          description: 분석 신뢰도

    # ==========================================
    # Financial Impact Schemas
    # ==========================================
...
    FinancialImpactResponse:
      type: object
      description: Hazard별 재무 영향 (AAL)
      properties:
        siteId:
          type: string
          format: uuid
        analyzedAt:
          type: string
          format: date-time
        hazardType:
          type: string
          example: 폭염
          description: 조회한 위험요인 (전체인 경우 null)
        assetValue:
          type: number
          format: double
          description: 자산 가치 (KRW)
          example: 10000000000
        financialImpacts:
          type: array
          description: Hazard별 재무 영향
          items:
            $ref: '#/components/schemas/FinancialImpactItem'

    # ==========================================
    # Vulnerability Schemas
    # ==========================================
...
    VulnerabilityResponse:
      type: object
      properties:
        siteId:
          type: string
          format: uuid
        totalScore:
          type: integer
          minimum: 0
          maximum: 100
          description: 종합 취약성 점수
        overallLevel:
          type: string
          enum: [낮음, 보통, 높음, 심각]
        siteInfo:
          type: object
          description: 사업장 정보
          properties:
            buildingAge:
              type: integer
            buildingType:
              type: string
            seismicDesign:
              type: boolean
            floorArea:
              type: number
        details:
          type: array
          description: Hazard별 취약성 상세
          items:
            $ref: '#/components/schemas/VulnerabilityItem'
        priorityActions:
          type: array
          description: 우선 조치 사항
          items:
            type: object
            properties:
              action:
                type: string
              priority:
                type: string
                enum: [low, medium, high, critical]
              estimatedCost:
                type: number

    # ==========================================
    # Total Analysis Response
    # ==========================================
    AnalysisTotalResponse:
      type: object
      description: 특정 Hazard 기준 통합 분석 결과
      properties:
        siteId:
          type: string
          format: uuid
        hazardType:
          type: string
          example: 폭염
        overview:
          $ref: '#/components/schemas/AnalysisOverviewResponse'
        pastEvents:
          $ref: '#/components/schemas/PastEventsResponse'
        sspProjection:
          $ref: '#/components/schemas/SSPProjectionResponse'
        financialImpact:
          $ref: '#/components/schemas/FinancialImpactResponse'
        vulnerability:
          $ref: '#/components/schemas/VulnerabilityResponse'
        executiveSummary:
          type: string
          description: AI가 생성한 경영진 요약 리포트
        generatedAt:
          type: string
          format: date-time

    # ==========================================
    # Simulation Schemas
    # ==========================================
...
    ReportGenerationRequest:
      type: object
      required: [scope, reportType]
      properties:
        scope:
          type: string
          enum: [single, all]
          description: >
            - single: 단일 사업장 리포트
            - all: 전체 사업장 통합 리포트
        siteId:
          type: string
          format: uuid
          description: 사업장 ID (scope=single일 때 필수)
        siteIds:
          type: array
          items:
            type: string
            format: uuid
          description: 포함할 사업장 목록 (scope=all일 때 선택, 미지정시 전체)
        reportType:
          type: string
          enum: [summary, full, governance]
          description: >
            - summary: 요약 리포트 (5-10페이지)
            - full: 전체 리포트 (30-50페이지)
            - governance: 거버넌스 리포트 (TCFD 기반)
        includeCharts:
          type: boolean
          default: true
          description: 차트 포함 여부
        language:
          type: string
          enum: [ko, en]
          default: ko
          description: 리포트 언어
        customSections:
          type: array
          description: 포함할 커스텀 섹션
          items:
            type: string
          example: [executive_summary, risk_matrix, action_plan]
    ReportGenerationStatus:
      type: object
      properties:
        reportId:
          type: string
          format: uuid
        scope:
          type: string
          enum: [single, all]
        siteId:
          type: string
          format: uuid
          description: scope=single일 때만 존재
        siteIds:
          type: array
          items:
            type: string
            format: uuid
          description: scope=all일 때 포함된 사업장 목록
        status:
          type: string
          enum: [queued, analyzing, generating, completed, failed]
          description: 리포트 생성 상태
        progress:
          type: integer
          minimum: 0
          maximum: 100
          description: 진행률 (%)
        currentStep:
          type: string
          example: LLM으로 요약 생성 중
          description: 현재 진행 단계
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        estimatedCompletionTime:
          type: string
          format: date-time
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
...
    ReportDownloadInfo:
      type: object
      properties:
        reportId:
          type: string
          format: uuid
        downloadUrl:
          type: string
          description: 리포트 다운로드 URL (임시 URL)
        format:
          type: string
          enum: [pdf, docx, json]
        fileSize:
          type: integer
          description: 파일 크기 (bytes)
        expiresAt:
          type: string
          format: date-time
          description: URL 만료 시간

    # ==========================================
    # Meta Schemas
    # ==========================================
    HazardType:
      type: object
      properties:
        code:
          type: string
          example: extreme_heat
        name:
          type: string
          example: 폭염
        nameEn:
          type: string
          example: Extreme Heat
        category:
          type: string
          enum: [temperature, water, wind, other]
          example: temperature
        description:
          type: string
          example: 일 최고기온 33도 이상이 2일 이상 지속되는 현상
        availableMetrics:
          type: array
          description: 분석 가능한 지표
          items:
            type: string
          example: [온도, 빈도, 지속시간]
paths:
...
  /api/v1/analysis/start:
    post:
      tags: [Analysis]
      summary: AI 리스크 분석 시작
      description: >
        AI Agent를 통해 사업장의 기후 물리적 리스크를 분석합니다.
        비동기로 처리되며 jobId를 반환합니다.
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartAnalysisRequest'
      responses:
        '202':
          description: 분석 작업이 시작됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJobStatus'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 이미 실행 중인 작업이 있음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/analysis/status/{jobId}:
    get:
      tags: [Analysis]
      summary: 분석 작업 상태 조회
      description: 실행 중인 분석 작업의 진행 상태를 조회합니다.
      security:
        - apiKeyAuth: []
      parameters:
        - in: path
          name: jobId
          required: true
          schema:
            type: string
            format: uuid
          description: 작업 ID
      responses:
        '200':
          description: 작업 상태
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJobStatus'
        '404':
          description: 작업을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/analysis/{siteId}/overview:
    get:
      tags: [Analysis]
      summary: 물리적 리스크 개요 조회
      description: 사업장의 물리적 리스크 분석 개요를 반환합니다.
      security:
        - apiKeyAuth: []
      parameters:
        - in: path
          name: siteId
          required: true
          schema:
            type: string
            format: uuid
          description: 사업장 ID
      responses:
        '200':
          description: 분석 개요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisOverviewResponse'
        '404':
          description: 분석 결과를 찾을 수 없음 (분석을 먼저 실행해야 함)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/analysis/{siteId}/physical-risk-scores:
    get:
      tags: [Analysis]
      summary: 물리적 리스크 점수 조회
      description: Hazard별 물리적 리스크 점수를 조회합니다.
      security:
        - apiKeyAuth: []
      parameters:
        - in: path
          name: siteId
          required: true
          schema:
            type: string
            format: uuid
          description: 사업장 ID
        - in: query
          name: hazardType
          schema:
            type: string
            enum: [폭염, 한파, 태풍, 홍수, 가뭄, 산불, 해안침수, 도시침수, 물부족]
          description: 특정 Hazard만 조회 (미지정시 전체)
      responses:
        '200':
          description: 물리적 리스크 점수
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhysicalRiskScoreResponse'
        '404':
          description: 분석 결과를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/analysis/{siteId}/past-events:
    get:
      tags: [Analysis]
      summary: 과거 재난 이력 조회
      description: 사업장 인근의 과거 재난 이력을 분석하여 반환합니다.
      security:
        - apiKeyAuth: []
      parameters:
        - in: path
          name: siteId
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: hazardType
          schema:
            type: string
            enum: [폭염, 한파, 태풍, 홍수, 가뭄, 산불, 해안침수, 도시침수, 물부족]
          description: 특정 Hazard만 조회 (미지정시 전체)
        - in: query
          name: startYear
          schema:
            type: integer
          description: 시작 연도
        - in: query
          name: endYear
          schema:
            type: integer
          description: 종료 연도
      responses:
        '200':
          description: 과거 재난 이력
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PastEventsResponse'
        '404':
          description: 분석 결과를 찾을 수 없음
  /api/v1/analysis/{siteId}/ssp:
    get:
      tags: [Analysis]
      summary: SSP 시나리오별 리스크 전망
      description: SSP 시나리오에 따른 미래 리스크 전망을 반환합니다.
      security:
        - apiKeyAuth: []
      parameters:
        - in: path
          name: siteId
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: hazardType
          schema:
            type: string
          description: 특정 Hazard (미지정시 주요 Hazard)
        - in: query
          name: timeScale
          schema:
            type: string
            enum: [단기, 중기, 장기]
          description: 시간 척도
      responses:
        '200':
          description: SSP 리스크 전망
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SSPProjectionResponse'
        '404':
          description: 분석 결과를 찾을 수 없음
  /api/v1/analysis/{siteId}/financial-impacts:
    get:
      tags: [Analysis]
      summary: 재무 영향 조회 (AAL)
      description: Hazard별 재무 영향(AAL)을 조회합니다.
      security:
        - apiKeyAuth: []
      parameters:
        - in: path
          name: siteId
          required: true
          schema:
            type: string
            format: uuid
          description: 사업장 ID
        - in: query
          name: hazardType
          schema:
            type: string
            enum: [폭염, 한파, 태풍, 홍수, 가뭄, 산불, 해안침수, 도시침수, 물부족]
          description: 특정 Hazard만 조회 (미지정시 전체)
      responses:
        '200':
          description: 재무 영향 분석
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinancialImpactResponse'
        '404':
          description: 분석 결과를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/analysis/{siteId}/vulnerability:
    get:
      tags: [Analysis]
      summary: 취약성 분석
      description: 사업장의 기후 리스크에 대한 취약성을 분석하여 반환합니다.
      security:
        - apiKeyAuth: []
      parameters:
        - in: path
          name: siteId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 취약성 분석 결과
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VulnerabilityResponse'
        '404':
          description: 분석 결과를 찾을 수 없음
  /api/v1/analysis/{siteId}/total:
    get:
      tags: [Analysis]
      summary: 특정 Hazard 기준 통합 분석 결과
      description: >
        특정 Hazard에 대한 모든 분석 결과를 통합하여 반환합니다.
        (개요 + 과거이력 + SSP전망 + 재무영향 + 취약성)
      security:
        - apiKeyAuth: []
      parameters:
        - in: path
          name: siteId
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: hazardType
          required: true
          schema:
            type: string
          description: 위험요인 유형
        - in: query
          name: timeScale
          schema:
            type: string
            enum: [단기, 중기, 장기]
      responses:
        '200':
          description: 통합 분석 결과
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisTotalResponse'
        '404':
          description: 분석 결과를 찾을 수 없음

  # ==========================================
  # Simulation Endpoints
  # ==========================================
...
  /api/v1/reports/generate:
    post:
      tags: [Reports]
      summary: 리포트 생성 시작
      description: >
        LLM을 사용하여 사업장 리스크 분석 리포트를 생성합니다.
        분석 데이터를 기반으로 LLM이 자동으로 텍스트를 작성합니다.
        비동기로 처리되며 reportId를 반환합니다.
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportGenerationRequest'
      responses:
        '202':
          description: 리포트 생성 작업이 시작됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportGenerationStatus'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 사업장 또는 분석 결과를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/reports/{reportId}/status:
    get:
      tags: [Reports]
      summary: 리포트 생성 상태 조회
      description: 리포트 생성 작업의 진행 상태를 조회합니다.
      security:
        - apiKeyAuth: []
      parameters:
        - in: path
          name: reportId
          required: true
          schema:
            type: string
            format: uuid
          description: 리포트 ID
      responses:
        '200':
          description: 리포트 생성 상태
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportGenerationStatus'
        '404':
          description: 리포트를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/reports/{reportId}/content:
    get:
      tags: [Reports]
      summary: 리포트 컨텐츠 조회
      description: >
        LLM이 생성한 리포트 컨텐츠를 JSON 형태로 반환합니다.
        Spring Boot는 이 데이터를 사용하여 PDF/Excel로 변환할 수 있습니다.
      security:
        - apiKeyAuth: []
      parameters:
        - in: path
          name: reportId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 리포트 컨텐츠
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportContent'
        '404':
          description: 리포트를 찾을 수 없음 또는 아직 생성 중
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/reports/{reportId}/download:
    get:
      tags: [Reports]
      summary: 리포트 파일 다운로드
      description: >
        생성된 리포트 파일을 다운로드합니다.
        FastAPI가 PDF/DOCX를 직접 생성하는 경우 사용합니다.
      security:
        - apiKeyAuth: []
      parameters:
        - in: path
          name: reportId
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: format
          schema:
            type: string
            enum: [pdf, docx, json]
            default: pdf
          description: 다운로드 형식
      responses:
        '200':
          description: 리포트 파일 또는 다운로드 URL
          content:
            application/pdf:
              schema:
                type: string
                format: binary
            application/vnd.openxmlformats-officedocument.wordprocessingml.document:
              schema:
                type: string
                format: binary
            application/json:
              schema:
                $ref: '#/components/schemas/ReportDownloadInfo'
        '404':
          description: 리포트를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ==========================================
  # Meta Endpoints
  # ==========================================
  /api/v1/meta/hazards:
    get:
      tags: [Meta]
      summary: 지원하는 기후 리스크(위험요인) 목록
      description: AI Agent가 분석 가능한 기후 위험요인 목록을 반환합니다.
      security:
        - apiKeyAuth: []
      parameters:
        - in: query
          name: category
          schema:
            type: string
            enum: [temperature, water, wind, other]
          description: 카테고리별 필터링
      responses:
        '200':
          description: 위험요인 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HazardType'
...
security:
  - apiKeyAuth: []